Final touches

+++ public/js/scripts.js
@@ -1,13 +1,15 @@
 function deleteContact(id)
 {
-    $.ajax({
-        type: "DELETE",
-        url: "/manage?id=" + id,
-        success: function(data) {
-            var json = JSON.parse(data);
-            window.location = "/?deleted=" + json.success;
-        }
-    });
+    if (confirm("Are you sure you want to delete this contact?")) {
+        $.ajax({
+            type: "DELETE",
+            url: "/manage?id=" + id,
+            success: function(data) {
+                var json = JSON.parse(data);
+                window.location = "/?deleted=" + json.deleted;
+            }
+        });
+    }
 }
 
 $(document).ready(function() {
